refactor(nav): extract RecoveryService from ExperimentalControls
Move the gesture-end recovery policy (_maybeRecoverAtGestureEnd) into
recoveryService.js: after a camera-drag ends non-legit, ease back to the stored
legit pose (runner Door 1) or pop to the roof. Reads the sensor's legit snapshot,
hands the motion to ctx.runner.runRecovery, and reaches the pop-to-roof fallback
via ctx.transition. The mouseup router calls this._recovery.maybeRecoverAtGestureEnd().

Behaviour-preserving: 302/302 characterization tests green.
